Develop

+++ products/ASC.Files/Core/Services/DocumentService/DocumentServiceConnector.cs
@@ -53,6 +53,7 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
                                       string region,
                                       ThumbnailData thumbnail,
                                       SpreadsheetLayout spreadsheetLayout,
+                                      Options options,
                                       bool isAsync,
                                       bool toForm)
     {
@@ -70,6 +71,7 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
                 region,
                 thumbnail,
                 spreadsheetLayout,
+                options,
                 isAsync,
                 fileUtility.SignatureSecret,
                 clientFactory,
@@ -125,16 +127,26 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
         string scriptUrl = null;
         if (!string.IsNullOrEmpty(inputScript))
         {
-            using (var stream = new MemoryStream())
-            await using (var writer = new StreamWriter(stream))
+            if (System.IO.File.Exists(inputScript))
             {
-                await writer.WriteAsync(inputScript);
-                await writer.FlushAsync();
-                stream.Position = 0;
-                scriptUrl = await pathProvider.GetTempUrlAsync(stream, ".docbuilder");
+                await using (var stream = System.IO.File.OpenRead(inputScript))
+                {
+                    scriptUrl = await pathProvider.GetTempUrlAsync(stream, ".docbuilder");
+                }
             }
-            scriptUrl = await ReplaceCommunityAddressAsync(scriptUrl);
-            requestKey = null;
+            else
+            {
+                using (var stream = new MemoryStream())
+                await using (var writer = new StreamWriter(stream))
+                {
+                    await writer.WriteAsync(inputScript);
+                    await writer.FlushAsync();
+                    stream.Position = 0;
+                    scriptUrl = await pathProvider.GetTempUrlAsync(stream, ".docbuilder");
+                }
+            }
+            scriptUrl = ReplaceCommunityAddress(scriptUrl);
+            requestKey = scriptUrl;
         }
 
         logger.DebugDocServiceBuilderRequestKey(requestKey, isAsync);
@@ -164,7 +176,7 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
                 fileUtility,
                 filesLinkUtility.DocServiceCommandUrl,
                 CommandMethod.Version,
-                GenerateRevisionId(null),
+                null,
                 null,
                 null,
                 null,
@@ -250,10 +262,10 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
                 var toExtension = fileUtility.GetInternalExtension(fileExtension);
                 var url = pathProvider.GetEmptyFileUrl(fileExtension);
 
-                var fileUri = await ReplaceCommunityAddressAsync(url);
+                var fileUri = ReplaceCommunityAddress(url);
 
                 var key = GenerateRevisionId(Guid.NewGuid().ToString());
-                var uriTuple = await ASC.Files.Core.Helpers.DocumentService.GetConvertedUriAsync(fileUtility, filesLinkUtility.DocServiceConverterUrl, fileUri, fileExtension, toExtension, key, null, null, null, null, false, fileUtility.SignatureSecret, clientFactory, false);
+                var uriTuple = await ASC.Files.Core.Helpers.DocumentService.GetConvertedUriAsync(fileUtility, filesLinkUtility.DocServiceConverterUrl, fileUri, fileExtension, toExtension, key, null, null, null, null, null, false, fileUtility.SignatureSecret, clientFactory, false);
                 convertedFileUri = uriTuple.ConvertedDocumentUri;
             }
             catch (Exception ex)
@@ -308,7 +320,7 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
                 var storeTemplate = await globalStore.GetStoreTemplateAsync();
                 var scriptUri = await storeTemplate.GetUriAsync("", "test.docbuilder");
                 var scriptUrl = baseCommonLinkUtility.GetFullAbsolutePath(scriptUri.ToString());
-                scriptUrl = await ReplaceCommunityAddressAsync(scriptUrl);
+                scriptUrl = ReplaceCommunityAddress(scriptUrl);
 
                 await ASC.Files.Core.Helpers.DocumentService.DocbuilderRequestAsync(fileUtility, filesLinkUtility.DocServiceDocbuilderUrl, null, scriptUrl, false, fileUtility.SignatureSecret, clientFactory);
             }
@@ -321,7 +333,7 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
         }
     }
 
-    public async Task<string> ReplaceCommunityAddressAsync(string url)
+    public string ReplaceCommunityAddress(string url)
     {
         var docServicePortalUrl = filesLinkUtility.GetDocServicePortalUrl();
 
@@ -332,7 +344,7 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
 
         if (string.IsNullOrEmpty(docServicePortalUrl))
         {
-            var tenant = await tenantManager.GetCurrentTenantAsync();
+            var tenant = tenantManager.GetCurrentTenant();
             if (!tenantExtra.Saas
                 || string.IsNullOrEmpty(tenant.MappedDomain)
                 || !url.StartsWith("https://" + tenant.MappedDomain))
@@ -343,7 +355,12 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
             docServicePortalUrl = "https://" + tenant.GetTenantDomain(coreSettings, false);
         }
 
-        var uri = new UriBuilder(url);
+        if (!Uri.TryCreate(url, UriKind.Absolute, out var absoluteUri))
+        {
+            return url;
+        }
+
+        var uri = new UriBuilder(absoluteUri);
         if (new UriBuilder(baseCommonLinkUtility.ServerRootPath).Host != uri.Host)
         {
             return url;
